requirements.txt
numpy==1.12.0
scipy==0.19.0
Keras==2.0.1
biopython==1.68
scikit-learn==0.18.1
gensim==1.0.1
